Pass redis conf file to redis-server program in supervisord.conf
Previously, the redis-server program was started from supervisord.conf
without a configuration file specifying the data directory to use. This
resulted in redis "not able to persist to disk" errors when the mispzmq
script performed some redis commands. The error was due to the default
redis data directory, /etc/redis, not being writable by the user that
the redis process runs as, redis.

This commit passes the /etc/redis/redis.conf configuration file to the
redis-server process at startup. The configuration file already sets the
data directory to a directory that the redis user has write access to,
/var/lib/redis.