Ensure pyzmq and redis also installed with pip2 for Python 2
Previously, pyzmq and redis were pip installed using the default pip,
which was for Python 3. This resulted in the pyzmq and redis packages
not being found when the MISP server spawned the mispzmq script using
the default python, which is version 2. With this commit, pyzmq and
redis would be installed for both Python 2 (via pip2) and Python 3
(via pip3), which should allow the packages to be available to scripts
which use it from different Python versions.