Populate uninitialized data directories at container startup
Previously, configuration for ssl cert/key and the MISP web
application did not have volume mounts defined by default and
population of the mysql database required creating an ephemeral
container instance to run an initialization script once before
creating the full MISP app container.

With the changes in this commit, volumes would be available by
default for mysql, ssl, and the MISP application.
Initialization of the data for each volume is also done post
container startup so that the data can be populated into
mounted directories if not already present. Since this logic
would be run on each container startup, the one-time creation
of a container to initialize the mysql database would no
longer be required.