Cybox library version error fix
After a clean install if you log into MISP and navigate to Administration -> Server Settings -> Diagnostics in the STIX and Cybox libraries section you will see the error CYBOX library version...Incorrect CyBox version installed, found 2.1.0.17 expecting 2.1.0.18.dev0.  I reached out the the developers of MISP and they told me to run pip3 uninstall cybox to uninstall the pip managed version.  To test it out I did I clean install and SSH'd into the container then ran the command and the error went away.  Not sure where the best place to incorporate this is, but this was were I tested it and I successfully ran a clean build with it.